CLN: separate shg-pl-trans objects from main pyfibre package
