+++ pyfibre/gui/tests/test_file_display_pane.py
@@ -2,9 +2,10 @@ import os
 from unittest import TestCase
 
 from pyfibre.gui.file_display_pane import FileDisplayPane
-from pyfibre.shg_pl_trans.shg_pl_trans_reader import SHGPLTransReader
 from pyfibre.tests.fixtures import (
-    directory, test_shg_pl_trans_image_path, test_shg_image_path)
+    directory,
+    test_image_path)
+from pyfibre.tests.probe_classes.readers import ProbeMultiImageReader
 
 source_dir = os.path.dirname(os.path.realpath(__file__))
 pyfibre_dir = os.path.dirname(os.path.dirname(source_dir))
@@ -15,9 +16,9 @@ class TestFileDisplayPane(TestCase):
     def setUp(self):
 
         self.file_display = FileDisplayPane(
-            supported_readers={'SHG-PL-Trans': SHGPLTransReader()}
+            supported_readers={'Probe': ProbeMultiImageReader()}
         )
-        self.file_path = test_shg_pl_trans_image_path
+        self.file_path = test_image_path
 
     def test_add_file(self):
 
@@ -25,15 +26,13 @@ class TestFileDisplayPane(TestCase):
         self.assertEqual(1, len(self.file_display.file_table))
 
         table_row = self.file_display.file_table[0]
-        self.assertEqual(
-            directory + '/test-pyfibre',
-            table_row.name)
-        self.assertEqual('SHG-PL-Trans', table_row.tag)
+        self.assertEqual('probe-file', table_row.name)
+        self.assertEqual('Probe', table_row.tag)
         self.assertListEqual(
-            [test_shg_pl_trans_image_path],
+            [test_image_path],
             table_row.file_names)
 
-        self.file_display.add_files(test_shg_image_path)
+        self.file_display.add_files(test_image_path)
         self.assertEqual(1, len(self.file_display.file_table))
 
     def test_add_directory(self):
@@ -42,12 +41,10 @@ class TestFileDisplayPane(TestCase):
         self.assertEqual(1, len(self.file_display.file_table))
 
         table_row = self.file_display.file_table[0]
-        self.assertEqual(
-            directory + '/test-pyfibre',
-            table_row.name)
-        self.assertEqual('SHG-PL-Trans', table_row.tag)
+        self.assertEqual('probe-file', table_row.name)
+        self.assertEqual('Probe', table_row.tag)
         self.assertListEqual(
-            [test_shg_pl_trans_image_path],
+            [test_image_path],
             table_row.file_names)
 
     def test_remove_file(self):
@@ -65,7 +62,7 @@ class TestFileDisplayPane(TestCase):
     def test_filter_files(self):
 
         self.file_display.add_files(self.file_path)
-        self.file_display.filter_files('pyfibre')
+        self.file_display.filter_files('probe-file')
 
         self.assertEqual(1, len(self.file_display.file_table))
