+++ pyfibre/shg_pl_trans/tests/fixtures/test-pyfibre-shg-Stack.tif
