+++ pyfibre/shg_pl_trans/tests/test_shg_reader.py
@@ -2,9 +2,11 @@ import numpy as np
 from skimage.external.tifffile import TiffFile
 
 from pyfibre.core.base_multi_image_reader import WrongFileTypeError
-from pyfibre.tests.fixtures import (
-    directory, test_shg_image_path, test_shg_pl_trans_image_path)
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
+from pyfibre.shg_pl_trans.tests.fixtures import (
+    directory,
+    test_shg_image_path,
+    test_shg_pl_trans_image_path)
 
 from ..shg_reader import (
     get_fluoview_param,
